Update and rename polyfitRidgeModelSel.py to polyfitRidgeCV.py
